Delete model modal and moved the right page to pages
