+++ src/app/app-routing.module.ts
@@ -16,7 +16,7 @@ import { ProjectNewComponent } from './projects/components/project-new/project-n
 import { ProjectAddComponent } from './projects/components/project-add/project-add.component';
 import { RecordIDEComponent } from './record-ide/components/record-ide.component';
 import { ConfigComponent } from './config/components/config.component';
-import { ModelDownloadComponent } from './model-download/components/model-download/model-download.component';
+import { ModelDownloadComponent } from './model-download/pages/model-download/model-download.component';
 
 const routes: Routes = [
   {
