+++ src/app/model-download/components/model-download-component/model-download-component.component.html
@@ -104,6 +104,7 @@
                             </td>
                             <td class="whitespace-nowrap text-center px-3 py-2 text-sm text-gray-500">
                                 <label [for]="'model-download'+model.name"
+                                    (click)="currentModel = model"
                                     class="modal-button btn btn-sm btn-ghost px-0 inline-block"
                                     [ngClass]="!isManaged ? 'pointer-events-none cursor-not-allowed opacity-50' : '' ">
 
@@ -113,30 +114,6 @@
                                             d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16" />
                                     </svg>
                                 </label>
-                                <input type="checkbox"
-                                    [id]="'model-download'+model.name"
-                                    class="modal-toggle">
-                                <div class="modal">
-                                    <div class="modal-box text-black bg-white justify-center">
-                                        <div
-                                            class="flex flex-grow justify-center text-lg leading-6 text-gray-900 font-medium">
-                                            Warning </div>
-                                        <p class="mt-2 text-gray-500 text-sm">Are you sure you want to
-                                            delete
-                                            this model?</p>
-                                        <p class="text-gray-500 text-sm">This will delete all data
-                                            associated
-                                            with it!</p>
-                                        <div class="modal-action">
-                                            <label [for]="'model-download'+model.name"
-                                                class="bg-red-100 text-red-700 border border-red-400 text-xs font-semibold mr-2 px-4 py-2 rounded flex cursor-pointer hover:bg-red-200 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-red-500 inline-block"
-                                                (click)="deleteModel(model.name, model.revision)">Delete
-                                                </label>
-                                            <label [for]="'model-download'+model.name"
-                                                class="bg-white text-gray-700 text-xs font-semibold mr-4 px-4 py-2 rounded border border-gray-300 cursor-pointer hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500 inline-block">Close</label>
-                                        </div>
-                                    </div>
-                                </div>
                             </td>
                             
                         </tr>
@@ -222,3 +199,28 @@
     </div>
 </ng-template>
 </div>
+
+<input type="checkbox"
+    [id]="'model-download'+currentModel?.name"
+    class="modal-toggle">
+    <div class="modal">
+        <div class="modal-box text-black bg-white justify-center text-center">
+            <div
+                class="flex flex-grow justify-center text-lg leading-6 text-gray-900 font-medium">
+                Warning </div>
+            <p class="mt-2 text-gray-500 text-sm">Are you sure you want to
+                delete
+                this model?</p>
+            <p class="text-gray-500 text-sm">This will delete all data
+                associated
+                with it!</p>
+            <div class="modal-action">
+                <label [for]="'model-download'+currentModel?.name"
+                    class="bg-red-100 text-red-700 border border-red-400 text-xs font-semibold mr-2 px-4 py-2 rounded flex cursor-pointer hover:bg-red-200 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-red-500 inline-block"
+                    (click)="deleteModel(currentModel?.name, currentModel?.revision)">Delete
+                    </label>
+                <label [for]="'model-download'+currentModel?.name"
+                    class="bg-white text-gray-700 text-xs font-semibold mr-4 px-4 py-2 rounded border border-gray-300 cursor-pointer hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500 inline-block">Close</label>
+            </div>
+        </div>
+    </div>
