+++ src/app/model-download/components/model-download-component/model-download-component.component.ts
@@ -28,6 +28,7 @@ export class ModelDownloadComponentComponent implements OnInit {
   currentModelHandle: any;
   downloadedModels: any[];
   subscriptions$: Subscription[] = [];
+  currentModel: any;
 
   constructor(
     private activatedRoute: ActivatedRoute,
