+++ src/app/model-download/model-download.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { BaseModule } from '../base/base.module';
-import { ModelDownloadComponent } from './components/model-download/model-download.component';
+import { ModelDownloadComponent } from './pages/model-download/model-download.component';
 import { ModelDownloadComponentComponent } from './components/model-download-component/model-download-component.component';
